% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_cropped_scene.R
\name{render_point_cropped_scene}
\alias{render_point_cropped_scene}
\title{Create a cropped scene}
\usage{
render_point_cropped_scene(ray_image, datamat, position, position_next,
  window = 100, zscale = 5, zoom = 0.4, windowsize = c(500, 500))
}
\arguments{
\item{ray_image}{Rayshaded image created with sphere_shade, add_shadow, ...}

\item{datamat}{Original raster transformed as matrix with t(as.matrix(raster))}

\item{position}{dataframe of point coordinates in the {rayshader} system
as created using sf_proj_as_ray()}

\item{position_next}{dataframe of another point coordinates in the {rayshader}
system as created using sf_proj_as_ray()}

\item{window}{size of the square window in pixels around the position}

\item{zscale}{Ratio between Z and x/y. Adjust the zscale down to exaggerate
elevation features. Better to use the same for the entire workflow.}

\item{zoom}{Zoom factor. Reduce to zoom in.}

\item{windowsize}{Size of the rgl window}
}
\description{
Create a cropped scene
}
