% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_proj_as_ray.R
\name{sf_proj_as_ray}
\alias{sf_proj_as_ray}
\title{Modify sf data to plot over rayshader}
\usage{
sf_proj_as_ray(r, sf, z_pos = 1, zscale = 10, as_sf = TRUE,
  groups = TRUE, crop = TRUE)
}
\arguments{
\item{r}{original raster used in rayshader}

\item{sf}{sf object to add on plot (with or without z column)}

\item{z_pos}{z-position of the sf on the 3d map if not in dataset}

\item{zscale}{if z-column existed in the dataset}

\item{as_sf}{Logical. Return sf object or not}

\item{groups}{Logical. Whether to group coordinates in blocks according to original structure.}

\item{crop}{Whether to crop sf object to extent of raster}
}
\description{
Modify sf data to plot over rayshader
}
